requirements.txt
PyYAML==3.12
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
idna==2.7
pip==10.0.1
requests==2.19.1
setuptools==39.1.0
urllib3==1.23